# specs/001-rag-chatbot/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RAG Chatbot API
  description: |
    API for the book-grounded RAG chatbot with personalization and translation features.
    All chat responses are derived strictly from indexed book content.
  version: 1.0.0
  contact:
    name: RAG Chatbot Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourbook.com
    description: Production server
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/preferences:
    put:
      tags: [Auth]
      summary: Update user preferences
      operationId: updateUserPreferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
          description: JWT token
        token_type:
          type: string
          default: bearer
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    UserProfileResponse:
      type: object
      required: [user, preferences]
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      required: [experience_level, preferred_language, chapters_read]
      properties:
        experience_level:
          type: string
          enum: [beginner, intermediate, advanced]
        preferred_language:
          type: string
          enum: [en, ur]
        chapters_read:
          type: array
          items:
            type: string
          description: List of chapter slugs the user has read
    UpdatePreferencesRequest:
      type: object
      properties:
        experience_level:
          type: string
          enum: [beginner, intermediate, advanced]
        preferred_language:
          type: string
          enum: [en, ur]
        chapters_read:
          type: array
          items:
            type: string
    UserPreferencesResponse:
      type: object
      required: [preferences]
      properties:
        preferences:
          $ref: '#/components/schemas/UserPreferences'

    # Errors
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - NOT_FOUND
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Additional error details (dev mode only)